feat(authentication): kv flow introduction for authentication  (#12611)
