+++ crates/diesel_models/src/query/authentication.rs
@@ -2,10 +2,8 @@ use diesel::{associations::HasTable, BoolExpressionMethods, ExpressionMethods};
 
 use super::generics;
 use crate::{
-    authentication::{
-        Authentication, AuthenticationNew, AuthenticationUpdate, AuthenticationUpdateInternal,
-    },
-    errors,
+    authentication::{Authentication, AuthenticationNew, AuthenticationUpdateInternal},
+    errors, kv,
     schema::authentication::dsl,
     PgPooledConn, StorageResult,
 };
@@ -14,14 +12,21 @@ impl AuthenticationNew {
     pub async fn insert(self, conn: &PgPooledConn) -> StorageResult<Authentication> {
         generics::generic_insert(conn, self).await
     }
+
+    pub async fn generate_drainer_insert_query(
+        self,
+        conn: &mut PgPooledConn,
+    ) -> StorageResult<kv::SerializableQuery> {
+        kv::generate_insert_query(conn, self).await
+    }
 }
 
 impl Authentication {
     pub async fn update_by_merchant_id_authentication_id(
         conn: &PgPooledConn,
-        merchant_id: common_utils::id_type::MerchantId,
-        authentication_id: common_utils::id_type::AuthenticationId,
-        authorization_update: AuthenticationUpdate,
+        merchant_id: &common_utils::id_type::MerchantId,
+        authentication_id: &common_utils::id_type::AuthenticationId,
+        authentication_update: AuthenticationUpdateInternal,
     ) -> StorageResult<Self> {
         match generics::generic_update_with_unique_predicate_get_result::<
             <Self as HasTable>::Table,
@@ -33,7 +38,7 @@ impl Authentication {
             dsl::merchant_id
                 .eq(merchant_id.to_owned())
                 .and(dsl::authentication_id.eq(authentication_id.to_owned())),
-            AuthenticationUpdateInternal::from(authorization_update),
+            authentication_update,
         )
         .await
         {
@@ -80,6 +85,24 @@ impl Authentication {
             dsl::merchant_id
                 .eq(merchant_id.to_owned())
                 .and(dsl::connector_authentication_id.eq(connector_authentication_id.to_owned())),
+        )
+        .await
+    }
+}
+
+impl AuthenticationUpdateInternal {
+    pub async fn generate_drainer_update_query(
+        self,
+        conn: &mut PgPooledConn,
+        merchant_id: common_utils::id_type::MerchantId,
+        authentication_id: common_utils::id_type::AuthenticationId,
+    ) -> StorageResult<kv::SerializableQuery> {
+        kv::generate_update_query_with_predicate::<<Authentication as HasTable>::Table, _, _>(
+            conn,
+            dsl::merchant_id
+                .eq(merchant_id)
+                .and(dsl::authentication_id.eq(authentication_id)),
+            self,
         )
         .await
     }
