+++ crates/router/src/core/authentication/transformers.rs
@@ -367,6 +367,7 @@ impl
             merchant_country_code: None,
             processor_merchant_id: None,
             created_by: None,
+            updated_by: None,
         }
     }
 }
@@ -491,6 +492,7 @@ impl
             merchant_country_code: None,
             processor_merchant_id: None,
             created_by: None,
+            updated_by: None,
         }
     }
 }
