+++ crates/router/src/core/payments/operations/payment_response.rs
@@ -2877,6 +2877,7 @@ async fn payment_response_update_tracker<F: Clone, T: types::Capturable>(
             Some(mut authentication_store) => {
                 let authentication_update = hyperswitch_domain_models::authentication::AuthenticationUpdate::PostAuthorizationUpdate {
                 authentication_lifecycle_status: enums::AuthenticationLifecycleStatus::Used,
+                updated_by: processor.get_account().storage_scheme.to_string(),
             };
                 let updated_authentication = state
                     .store
@@ -2885,6 +2886,7 @@ async fn payment_response_update_tracker<F: Clone, T: types::Capturable>(
                         authentication_update,
                         processor.get_key_store(),
                         &key_manager_state,
+                        processor.get_account().storage_scheme,
                     )
                     .await
                     .to_not_found_response(errors::ApiErrorResponse::PaymentNotFound)?;
@@ -2893,7 +2895,7 @@ async fn payment_response_update_tracker<F: Clone, T: types::Capturable>(
             }
             None => None,
         };
-    };
+    }
 
     let amount_captured = get_total_amount_captured(
         &router_data.request,
