+++ crates/router/src/core/payments/operations/payment_status.rs
@@ -485,6 +485,7 @@ async fn get_tracker_for_sync<
                     authentication_id,
                     platform.get_processor().get_key_store(),
                     key_manager_state,
+                    platform.get_processor().get_account().storage_scheme,
                 )
                 .await
                 .to_not_found_response(errors::ApiErrorResponse::InternalServerError)
