+++ crates/storage_impl/src/lib.rs
@@ -560,6 +560,19 @@ impl UniqueConstraints for diesel_models::Mandate {
     }
 }
 
+impl UniqueConstraints for diesel_models::authentication::Authentication {
+    fn unique_constraints(&self) -> Vec<String> {
+        // Mirror the DB's only uniqueness: the `authentication_id` primary key.
+        vec![format!(
+            "authentication_{}",
+            self.authentication_id.get_string_repr()
+        )]
+    }
+    fn table_name(&self) -> &str {
+        "Authentication"
+    }
+}
+
 #[cfg(feature = "v1")]
 impl UniqueConstraints for diesel_models::Customer {
     fn unique_constraints(&self) -> Vec<String> {
