+++ crates/storage_impl/src/redis/kv_store.rs
@@ -46,6 +46,9 @@ pub enum PartitionKey<'a> {
         merchant_id: &'a common_utils::id_type::MerchantId,
         mandate_id: &'a str,
     },
+    AuthenticationId {
+        authentication_id: &'a common_utils::id_type::AuthenticationId,
+    },
     #[cfg(feature = "v2")]
     GlobalId {
         id: &'a str,
@@ -99,6 +102,9 @@ impl std::fmt::Display for PartitionKey<'_> {
                 "mid_{}_mandate_{mandate_id}",
                 merchant_id.get_string_repr()
             )),
+            PartitionKey::AuthenticationId { authentication_id } => {
+                f.write_str(authentication_id.get_string_repr())
+            }
 
             #[cfg(feature = "v2")]
             PartitionKey::GlobalId { id } => f.write_str(&format!("global_cust_{id}")),
